# apps/web/openapi/shipgate.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shipgate API",
    "description": "Shipgate HTTP API.",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "shipgateSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-shipgate_session",
        "description": "HttpOnly Shipgate browser session cookie."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/changes/{changeId}/exclusion": {
      "put": {
        "operationId": "excludeProjectChangeFromCandidate",
        "summary": "Exclude one change from the active draft candidate",
        "tags": ["Release planning"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 4000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "changeId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            },
            "in": "header",
            "name": "x-csrf-token",
            "required": false
          }
        ],
        "security": [
          {
            "shipgateSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "candidateId",
                    "candidateVersion",
                    "changeId",
                    "excluded",
                    "evaluationRequestId"
                  ],
                  "properties": {
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": ["recorded"]
                        },
                        {
                          "type": "string",
                          "enum": ["already_applied"]
                        }
                      ]
                    },
                    "candidateId": {
                      "type": "string"
                    },
                    "candidateVersion": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "changeId": {
                      "type": "string"
                    },
                    "excluded": {
                      "type": "boolean"
                    },
                    "evaluationRequestId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Common error response returned by the Shipgate API.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "example": "VALIDATION_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "details": {}
                  },
                  "additionalProperties": false,
                  "title": "ApiError",
                  "description": "Common error response returned by the Shipgate API."
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "restoreProjectChangeToCandidate",
        "summary": "Restore one explicitly excluded change to the active draft candidate",
        "tags": ["Release planning"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "changeId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            },
            "in": "header",
            "name": "x-csrf-token",
            "required": false
          }
        ],
        "security": [
          {
            "shipgateSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "candidateId",
                    "candidateVersion",
                    "changeId",
                    "excluded",
                    "evaluationRequestId"
                  ],
                  "properties": {
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": ["recorded"]
                        },
                        {
                          "type": "string",
                          "enum": ["already_applied"]
                        }
                      ]
                    },
                    "candidateId": {
                      "type": "string"
                    },
                    "candidateVersion": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "changeId": {
                      "type": "string"
                    },
                    "excluded": {
                      "type": "boolean"
                    },
                    "evaluationRequestId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Common error response returned by the Shipgate API.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "example": "VALIDATION_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "details": {}
                  },
                  "additionalProperties": false,
                  "title": "ApiError",
                  "description": "Common error response returned by the Shipgate API."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Shipgate origin"
    }
  ],
...
}